refatoração do codigo e funcionando a criação de usuarios
