+++ package.json
@@ -10,9 +10,12 @@
     "dev": "nodemon debugging/apolloServer.js"
   },
   "dependencies": {
+    "@graphql-tools/load-files": "^6.5.3",
+    "@graphql-tools/merge": "^8.2.1",
     "@graphql-tools/schema": "^8.3.1",
     "@prisma/client": "^3.8.1",
     "apollo-server": "^3.6.1",
+    "apollo-server-express": "^3.6.1",
     "cors": "^2.8.5",
     "express": "^4.17.2",
     "express-graphql": "^0.12.0",
