+++ src/app.js
@@ -1,20 +1,9 @@
-const express = require('express');
-const { graphqlHTTP } = require('express-graphql');
-const cors = require('cors');
-const { schema } = require('./Schema');
+const { ApolloServer } = require('apollo-server');
+const { typeDefs, resolvers } = require('./graphql');
 
-const main = async () => {
-  const app = express();
-  const PORT = 4000;
-  app.use(cors());
-  app.use(express.json());
+const server = new ApolloServer({
+  typeDefs,
+  resolvers,
+});
 
-  app.use('/graphql', graphqlHTTP({
-    schema,
-    graphiql: true,
-  }));
-
-  app.listen(PORT, () => console.log(`http://localhost:${PORT}`));
-};
-
-main().catch((error) => console.log(error));
+server.listen().then(({ url }) => console.log(`server running ${url}`));
