+++ src/graphql/modules/address/resolvers.js
@@ -1,0 +1,9 @@
+const { PrismaClient } = require('@prisma/client');
+
+module.exports = {
+  Query: {
+    addresses() {
+      return new PrismaClient().address.findMany();
+    },
+  },
+};
