+++ src/graphql/modules/address/schema.gql
@@ -1,0 +1,15 @@
+ type Address {
+  id: Int
+  street: String
+  number: Int
+  complement: String
+  city: String
+  state: String
+  country: String
+  zipCode: String
+}
+
+type Query {
+  address(id: Int): Address
+  addresses: [Address]
+}
